pyproject.toml
[project]
name = "menestrello"
license = "MIT"
description = "An AI powered interactive storyteller."
authors = [
    {name = "Matteo Angeloni", email = "mattange@gmail.com>"}
]
readme = "README.md"
readme_content_type = "text/markdown"
keywords = ["menestrello","story","interactive","AI","storyteller","children","education"]
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "google-cloud-texttospeech (>=2.25.1,<3.0.0)",
    "openai (>=1.69.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "treelib (>=1.7.1,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "playsound3 (>=3.2.3,<4.0.0)",
    "pynput (>=1.8.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]
requires-python = ">=3.13,<4.0"

[project.scripts]
menestrello = "menestrello.main:main"
generate_static_audio_eng = "menestrello.generate_static_audio_eng:main"

[tool.project.urls]
homepage = "https://github.com/mattange/menestrello"
repository = "https://github.com/mattange/menestrello"

[tool.poetry]
version = "0.1.0"
requires-poetry = ">=2.0"
packages = [{include = "menestrello", from="src"}]

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^9.21.0"
ipykernel = "^6.29.5"

[tool.poetry.group.tests.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
allow_zero_version = false
build_command = "poetry build"
commit_parser = "emoji"
version_toml = [
    "pyproject.toml:tool.poetry.version",  # Implied Default: Number format
]
version_variables = [
    "src/menestrello/__init__.py:__version__",
]
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false